requirements.txt
ultralytics>=8.0.0
flask>=3.0.0
opencv-python>=4.8.0
pillow>=10.2.0
numpy>=1.24.0
kaggle>=1.5.0
kagglehub>=0.2.0
torch>=2.0.0
timm>=0.9.0
